xpu: Use float32 instead of float64 for XPU compatibility

+++ projects/mmdet3d_plugin/models/detectors/bevdepth.py
@@ -71,10 +71,10 @@ class BEVDepth(BEVDet):
 
         # calculate the transformation from adj sensor to key ego
         keyego2global = ego2globals[:, 0,  ...].unsqueeze(1)    # (B, 1, 4, 4)
-        global2keyego = torch.inverse(keyego2global.double())   # (B, 1, 4, 4)
+        # Use float32 instead of float64 for XPU compatibility
+        global2keyego = torch.inverse(keyego2global.float())   # (B, 1, 4, 4)
         sensor2keyegos = \
-            global2keyego @ ego2globals.double() @ sensor2egos.double()     # (B, N_views, 4, 4)
-        sensor2keyegos = sensor2keyegos.float()
+            global2keyego @ ego2globals.float() @ sensor2egos.float()     # (B, N_views, 4, 4)
 
         return [imgs, sensor2keyegos, ego2globals, intrins,
                 post_rots, post_trans, bda]
